+++ Interpreter.Tests/Unit/IR/Types/ObjectTypeTests.cs
@@ -2,27 +2,10 @@ using System.Collections.Generic;
 using Interpreter.Lib.IR.CheckSemantics.Types;
 using Xunit;
 
-namespace Interpreter.Tests.Unit.IR
+namespace Interpreter.Tests.Unit.IR.Types
 {
-    public class TypeTests
+    public class ObjectTypeTests
     {
-        [Fact]
-        public void TypeEqualityTest()
-        {
-            var number = new Type("number");
-            var arrayOfNumbers = new ArrayType(number);
-            Assert.False(arrayOfNumbers.Equals(number));
-            Assert.False(number.Equals(arrayOfNumbers));
-        }
-
-        [Fact]
-        public void TypeStringRepresentationTest()
-        {
-            var matrix = new ArrayType(new ArrayType(new Type("number")));
-            
-            Assert.Equal("number[][]", matrix.ToString());
-        }
-
         [Fact]
         public void ObjectTypeEqualityTest()
         {
@@ -62,38 +45,14 @@ namespace Interpreter.Tests.Unit.IR
             Assert.NotEqual(p3d1, p3d2);
             Assert.NotEqual(p3d2, p2d1);
         }
-
+        
         [Fact]
-        public void NullTests()
-        {
-            var number = new Type("number");
-            // ReSharper disable once SuspiciousTypeConversion.Global
-            Assert.True(new NullType().Equals(new NullableType(number)));
-        }
-
-        [Fact]
-        public void TypeWrappingTest()
-        {
-            var str = new Type("string");
-            str = new NullableType(str);
-            str = new ArrayType(str);
-            Assert.Equal("string?[]", str.ToString());
-        }
-
-        [Fact]
-        public void DefaultValueTest()
-        {
-            Assert.Null(TypeUtils.GetDefaultValue(new NullableType(new Any())));
-            Assert.Null(TypeUtils.GetDefaultValue(new NullType()));
-            Assert.Null(TypeUtils.GetDefaultValue(new ObjectType(new List<PropertyType>())));
-        }
-
-        [Fact]
-        public void RecursiveTypeTest()
+        public void RecursiveTypeReferenceResolvingTest()
         {
             var number = new Type("number");
             var array = new ArrayType(new Type("self"));
             var method = new FunctionType(number, new List<Type> { new("self") });
+            var nullable = new NullableType(new Type("self"));
             var linkedListType = new ObjectType(
                 new List<PropertyType>
                 {
@@ -103,14 +62,16 @@ namespace Interpreter.Tests.Unit.IR
                         new("next", new Type("self"))
                     })),
                     new("children", array),
+                    new("parent", nullable),
                     new("compare", method)
                 }
             );
-            
+
             linkedListType.ResolveSelfReferences("self");
             
             Assert.Equal(linkedListType, ((ObjectType)linkedListType["wrapped"])["next"]);
             Assert.Equal(linkedListType, array.Type);
+            Assert.Equal(linkedListType, nullable.Type);
             Assert.Equal(linkedListType, method.Arguments[0]);
         }
 
@@ -137,6 +98,7 @@ namespace Interpreter.Tests.Unit.IR
             var number = new Type("number");
             var array = new ArrayType(new Type("self"));
             var method = new FunctionType(number, new List<Type> { new("self") });
+            var nullable = new NullableType(new Type("self"));
             var linkedListType = new ObjectType(
                 new List<PropertyType>
                 {
@@ -146,12 +108,43 @@ namespace Interpreter.Tests.Unit.IR
                         new("next", new Type("self"))
                     })),
                     new("children", array),
+                    new("parent", nullable),
                     new("compare", method)
                 }
             );
             
             linkedListType.ResolveSelfReferences("self");
             Assert.Contains("@this", linkedListType.ToString());
+        }
+
+        [Fact]
+        public void SerializationOfTypeWithRecursivePropertyTest()
+        {
+            var nodeType = new ObjectType(
+                new List<PropertyType>
+                {
+                    new("data", new Type("number")),
+                    new("next", new Type("self"))
+                }
+            )
+            {
+                Recursive = true
+            };
+            nodeType.ResolveSelfReferences("self");
+
+            var linkedListType = new ObjectType(
+                new List<PropertyType>
+                {
+                    new("head", nodeType),
+                    new("append", new FunctionType(
+                            new Type("void"),
+                            new List<Type> { nodeType }
+                        )
+                    )
+                }
+            );
+
+            Assert.Contains("head: head;", linkedListType.ToString());
         }
     }
 }
